compile/verify now uses arduino-cli daemon
The build options are now passed through GRPC.
There should be no need to pass library or hardware paths anymore.
By default the daemon compiles and export the hex in the sketch folder
but this is disallowed by the IDE unless the user explicitly run
the "Export .hex" command.

All the business logic to compose board properties and command line
for the build is now embedded in the daemon and has been removed from
the IDE. \o/